refactor(mcp): extract evolve_restore tool module
