+++ openspec/changes/refactor-mcp-evolve-restore-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-evolve-restore-tool-module
+
+Refactor: extract MCP evolve_restore tool implementation into a dedicated module
