+++ src/mcp/tools.rs
@@ -20,6 +20,7 @@ use crate::user_error::UserError;
 use super::confirm::{CONFIRM_TOKEN_TTL, ConfirmTokenBinding};
 use super::{AgentpackMcp, confirm};
 
+mod evolve_restore;
 mod explain;
 
 pub(super) const TOOLS_INSTRUCTIONS: &str = "Agentpack MCP server (stdio). Tools: plan, diff, preview, status, doctor, deploy, deploy_apply, rollback, evolve_propose, evolve_restore, explain.";
@@ -526,86 +527,7 @@ async fn call_rollback_in_process(
 async fn call_evolve_restore_in_process(
     args: EvolveRestoreArgs,
 ) -> anyhow::Result<(String, serde_json::Value)> {
-    tokio::task::spawn_blocking(move || {
-        let repo_override = args.common.repo.as_ref().map(std::path::PathBuf::from);
-        let profile = args.common.profile.as_deref().unwrap_or("default");
-        let target = args.common.target.as_deref().unwrap_or("all");
-        let machine_override = args.common.machine.as_deref();
-        let dry_run = args.common.dry_run.unwrap_or(false);
-
-        let engine = match crate::engine::Engine::load(repo_override.as_deref(), machine_override) {
-            Ok(v) => v,
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error("evolve.restore", &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                return Ok((text, envelope));
-            }
-        };
-
-        let result = crate::handlers::evolve::evolve_restore_in(
-            &engine,
-            profile,
-            target,
-            args.module_id.as_deref(),
-            dry_run,
-            args.yes,
-            true,
-        );
-
-        let (text, envelope) = match result {
-            Ok(crate::handlers::evolve::EvolveRestoreOutcome::Done(report)) => {
-                let data = crate::app::evolve_restore_json::evolve_restore_json_data(
-                    report.restored,
-                    report.summary,
-                    report.reason,
-                );
-                let mut envelope = crate::output::JsonEnvelope::ok("evolve.restore", data);
-                envelope.warnings = report.warnings;
-                let text = serde_json::to_string_pretty(&envelope)?;
-                let envelope = serde_json::to_value(&envelope)?;
-                (text, envelope)
-            }
-            Ok(crate::handlers::evolve::EvolveRestoreOutcome::NeedsConfirmation) => {
-                let err = UserError::confirm_required("evolve restore");
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error("evolve.restore", &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                (text, envelope)
-            }
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                let envelope = envelope_error("evolve.restore", &code, &message, details);
-                let text = serde_json::to_string_pretty(&envelope)?;
-                (text, envelope)
-            }
-        };
-
-        Ok((text, envelope))
-    })
-    .await
-    .context("mcp evolve_restore handler task join")?
+    evolve_restore::call_evolve_restore_in_process(args).await
 }
 
 async fn call_evolve_propose_in_process(
